ft/ migration of kaamelott quiz into an activity and redesign of HomeFragment
