+++ RickAndMortyAPP/app/src/main/res/layout/activity_kaamelott.xml
@@ -1,0 +1,147 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/collectionFragment"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+
+    android:background="@color/colorPrimary"
+    tools:context=".View.Collection.list.CollectionFragment">
+
+    <TextView
+        android:id="@+id/tv_actual_score"
+        style="@style/TextViewStyleWhite"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="24dp"
+        android:layout_marginStart="4dp"
+        android:text="@string/actual_score"
+        android:visibility="invisible"
+        tools:visibility="visible"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        />
+
+    <TextView
+        android:id="@+id/tv_actual_turn"
+        style="@style/TextViewStyleWhite"
+        android:text="@string/actual_turn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="16dp"
+        android:layout_marginEnd="4dp"
+        android:visibility="invisible"
+        tools:visibility="visible"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="@id/tv_actual_score" />
+
+    <ScrollView
+        android:id="@+id/scrollView"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginBottom="16dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tv_actual_turn"
+        app:layout_constraintBottom_toTopOf="@id/btn_perso1">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="12dp"
+            android:orientation="vertical" >
+
+            <TextView
+                android:id="@+id/tv_citation"
+                style="@style/quote_style"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:textAlignment="center"
+                android:visibility="invisible"
+                tools:visibility="visible"
+                tools:text="@tools:sample/lorem/random"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"/>
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </ScrollView>
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btn_perso1"
+        android:layout_width="0dp"
+        android:layout_height="82dp"
+        android:layout_margin="4dp"
+        tools:text="Perceval"
+        android:visibility="invisible"
+        tools:visibility="visible"
+        android:backgroundTint="@color/colorPrimaryDark"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toStartOf="@id/btn_perso2"
+        app:layout_constraintBottom_toTopOf="@id/btn_perso3"
+        />
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btn_perso2"
+        android:layout_width="0dp"
+        android:layout_height="82dp"
+        android:layout_margin="4dp"
+        tools:text="Arthur"
+        android:visibility="invisible"
+        tools:visibility="visible"
+        android:backgroundTint="@color/colorPrimaryDark"
+        app:layout_constraintStart_toEndOf="@id/btn_perso1"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/btn_perso3"
+        />
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btn_perso3"
+        style="@style/btn_style"
+        android:layout_width="0dp"
+        android:layout_height="82dp"
+        android:layout_margin="4dp"
+        tools:text="Leodagan"
+        android:visibility="invisible"
+        tools:visibility="visible"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toStartOf="@id/btn_perso4"
+        app:layout_constraintBottom_toBottomOf="parent"
+        />
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btn_perso4"
+        style="@style/btn_style"
+        android:layout_width="0dp"
+        android:layout_height="82dp"
+        android:layout_margin="4dp"
+        tools:text="Yvain"
+        android:visibility="invisible"
+        tools:visibility="visible"
+        app:layout_constraintStart_toEndOf="@id/btn_perso3"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        />
+
+    <TextView
+        android:id="@+id/tv_game_over"
+        style="@style/tv_game_over"
+        android:background="@drawable/back_game_over"
+        android:padding="8dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/try_an_other_game_later"
+        tools:text="@string/try_an_other_game_later"
+        android:textAlignment="center"
+        android:visibility="gone"
+        tools:visibility="visible"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
